portal: rework UI, introduce bootstrap.
This commit reworks all html content. It introduces bootstrap. This
allows us to create a responsive UI. Moreover it helps to keep the
code clean.

Beside bootstrap, this commit introduces a set of related css and js
files (minified). To be able to use bootstrap 4 we ship our own
version of it instead of relying on the version which is already
shipped with debian. The minified file versions make lintian
cry. However, this should be fixed in following commits.

To ship the new files as well as our own css file we use a new
directory /usr/share/elephant-shed/static. Which is available through
/static location.

changes. Lines starting