Inject navbar into webpages using mod_substitute
The portal page was embedding the actual content into an iframe, with
the navigation bar in the top window. The downside was that the URL
displayed was always just "/", and pgAdmin4 didn't like being embedded.

Now we access all pages directly, and use mod_substitute to inject a
position:absolute div into all html pages, plus a CSS stylesheet.

Known issues: It doesn't work with shellinabox (injection works, but
content is later replaced again).

Navbar links are broken in Cockpit (JS in action to redirect all links
to the websocket).

Navbar is displayed twice in /pgbadger (the index_header subrequest is
rewritten as well).