Revert "Fix deprecated ansible include"
This reverts commit 0b63ccf38190119353672862cb0167b6cc4cf406.

Does not work with ansible 2.2.1 yet.

$ ansible-playbook -i inventory elephant-shed.yml
ERROR! no action detected in task. This often indicates a misspelled module name, or incorrect module path.

The error appears to have been in '/builds/dbteam/elephant-shed-github-sync/vagrant/elephant-shed.yml': line 18, column 5, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

  handlers:
  - import_tasks: handlers.yml
    ^ here