ansible: add security repository if its not present.
We want to add the debian security repository, if not already
present. This is required if some 3rd party package depends on
versions that are not included in stretch main but in stretch/updates
main.  We use lineinfile instead of apt_repository to not add it if
its already configured in /etc/apt/sources.list. apt_repository would
add a new entry within /etc/apt/sources.list.d.