Allow overriding the interval for long-running sql-exporter jobs
The prometheus-sql-exporter allows setting the interval between queries for
long-running jobs. We expose this setting in the config generator, by allowing
users to set an interval value on their queries.

For each cluster and each database, this creates one job for all zero-interval
queries (that are supposed to be instantaneous), and one job for each
non-zero-interval query (so that they can drift apart if their runtimes are
different).