Fixed "Column 'usename' must be type text (string)"
Added "and usename is not null" to ensure no null-value usernames are returned to the sql-exporter.
The SQL-Exporter fails if null-values are returned. In the event of an autovacuum run, the usename is 
null and the database is not ( hence ignoring the datname predicate ).

We could separately count background-processes, but as we already have maintenance-counters
there is little reason to have a fuzzy-version in here.