Remove no longer needed node-exporter option.
The option --collector.netdev.ignored-devices was set in order to ignore the
loopback device, however, this option has been removed in node-exporter 1.0 and
the exporter fails to start in this case.

On earlier versions, Debian patched node-exporter to ignore the loopback device
anyway, so setting is was redundant. Hence, remove it, fixing #26.