Dashboard: Update Connections panels.
This changes the area fill to 1 from 5 to be more in line with the other
Database-related panels. For the Connections by State panel, also add a
table-based legend, only show connections states that are present in the chosen
time range and override the colors to be more in line with their states.